% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{ggplot.lm}
\alias{ggplot.lm}
\title{ggplot.lm}
\usage{
\method{ggplot}{lm}(data, mapping, vars, ...)
}
\arguments{
\item{data}{The linear model object from \code{lm}}

\item{mapping}{Regular mapping, see \code{ggplot2} and \code{aes} for details.}

\item{vars}{A list of variable values used for prediction.}

\item{\ldots}{Additional arguments passed to \code{ggplot2}}
}
\value{
A ggplot class object
}
\description{
Method for plotting linear models.

Simple method to plot a linear model using ggplot 
along with 95% confidence intervals
}
\examples{
ggplot2::ggplot( 
  lm(mpg ~ hp * qsec, data = mtcars), 
  ggplot2::aes(hp, mpg, linetype = factor(qsec)), 
  vars = list( 
    hp = min(mtcars$hp):max(mtcars$hp), 
    qsec = round(mean(mtcars$qsec) + c(-1, 1) * sd(mtcars$qsec)), 1)) + 
  ggplot2::geom_ribbon(ggplot2::aes(ymin = LL, ymax = UL), alpha = .2) + 
  ggplot2::geom_line() + 
  ggplot2::theme_bw() 
}

