% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{meanPlot}
\alias{meanPlot}
\title{meanPlot}
\usage{
meanPlot(formula, d)
}
\arguments{
\item{formula}{A formula specifying the variable to be used on the 
y-axis and the variable to be used on the x-axis}

\item{d}{A data.frame class object containing the variables specified 
in the \code{formula}}
}
\value{
Called for the side effect of creating a plot
}
\description{
This is a simple function designed to facilitate plotting raw
data along with dots indicating the mean at each x-axis value.

Although this function can be used with any type of data that works 
with \code{plot}, it works best when x-axis values are discrete
so that there are several y-values at the same x-axis value so that
the mean of multiple values is taken.
}
\examples{
#example using meanPlot
meanPlot(mpg ~ factor(cyl), d = mtcars)
}
\author{
Oscar Garcia-Cabrejo
}
\keyword{plot}

