% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleData.R
\docType{data}
\name{sampleData}
\alias{sampleData}
\title{Numbers and Letters}
\format{An object of class \code{data.frame} with 10 rows and 2 columns.}
\usage{
sampleData
}
\description{
A sample data set containing 10 numbers and letters with two variables:
 
 \itemize{
   \item Num: A number.
   \item Letter: An upper case letter (A to J)
 
 }  
 
 @format A data frame with 10 rows and 2 variables
 @source Created as sample
}
\keyword{datasets}

