% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textplot.R
\docType{class}
\name{textplot-class}
\alias{textplot-class}
\alias{textplot}
\alias{show,textplot-method}
\alias{[,textplot-method}
\title{An S4 class to hold text and Cartesian coordinates for plotting}
\usage{
\S4method{show}{textplot}(object)

\S4method{[}{textplot}(x, i, j, drop)
}
\arguments{
\item{object}{The object to be shown}

\item{x}{the object to subset}

\item{i}{the rows to subset (optional)}

\item{j}{the columns to subset (optional)}

\item{drop}{should be missing}
}
\description{
A class designed to hold the data required to create a textplot
where character strings are plotted based on x and y coordinates.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: show method

\item \code{[}: extract method
}}

\section{Slots}{

\describe{
\item{\code{x}}{A numeric value with the x axis coordinates.}

\item{\code{y}}{A numeric value with the y axis coordinates.}

\item{\code{labels}}{A character string with the text to be plotted}
}}

