% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textplot.R
\docType{class}
\name{textplot-class}
\alias{[,textplot-method}
\alias{show,textplot-method}
\alias{textplot}
\alias{textplot-class}
\title{textplot}
\usage{
\S4method{show}{textplot}(object)

\S4method{[}{textplot}(x, i, j, drop)
}
\arguments{
\item{object}{The object to be shown}

\item{x}{the object to subset}

\item{i}{the rows to subset}

\item{j}{the columns to subset}

\item{drop}{should be missing}
}
\description{
An S4 class to hold test and cartesian coordinates for plotting

A class designed to hold the data required to create a textplot
where character strings are plotted based on x and y coordinates
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: show method

\item \code{[}: extract method
}}
\section{Slots}{

\describe{
\item{\code{x}}{A numeric value with the x coordinates}

\item{\code{y}}{A numeric value with the y coordinates}

\item{\code{labels}}{A character string wit hthe text to be plotted}
}}

