% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textplot.R
\docType{class}
\name{textplot-class}
\alias{textplot}
\alias{textplot-class}
\title{textplot}
\description{
An S4 class to hold test and cartesian coordinates for plotting

A class designed to hold the data required to create a textplot
where character strings are plotted based on x and y coordinates
}
\section{Slots}{

\describe{
\item{\code{x}}{A numeric value with the x coordinates}

\item{\code{y}}{A numeric value with the y coordinates}

\item{\code{labels}}{A character string wit hthe text to be plotted}
}}

